refactor: Add formatting to an error message from EnvFileNames class
